+++ tests/test_adjoint_jacobian.py
@@ -223,6 +223,32 @@ class TestAdjointJacobian:
         numeric_val = fn(qml.execute(tapes, dev, None))
         assert np.allclose(calculated_val, numeric_val, atol=tol, rtol=0)
 
+    @pytest.mark.skipif(
+        device_name != "lightning.qubit" or not ld._CPP_BINARY_AVAILABLE,
+        reason="N-controlled operations only implemented in lightning.qubit.",
+    )
+    @pytest.mark.parametrize("n_qubits", [1, 2, 3, 4])
+    @pytest.mark.parametrize("par", [-np.pi / 7, np.pi / 5, 2 * np.pi / 3])
+    def test_phaseshift_gradient(self, n_qubits, par, tol):
+        """Test that the gradient of the phaseshift gate matches the exact analytic formula."""
+        par = np.array(par)
+        dev = qml.device("lightning.qubit", wires=n_qubits)
+        init_state = np.zeros(2**n_qubits)
+        init_state[-2::] = np.array([1.0 / np.sqrt(2), 1.0 / np.sqrt(2)], requires_grad=False)
+
+        with qml.tape.QuantumTape() as tape:
+            qml.StatePrep(init_state, wires=range(n_qubits))
+            qml.ctrl(qml.PhaseShift(par, wires=n_qubits - 1), range(0, n_qubits - 1))
+            qml.expval(qml.PauliY(n_qubits - 1))
+
+        tape.trainable_params = {1}
+
+        exact = np.cos(par)
+        grad_A = dev.adjoint_jacobian(tape)
+
+        # different methods must agree
+        assert np.allclose(grad_A, exact, atol=tol, rtol=0)
+
     @pytest.mark.parametrize("par", [1, -2, 1.623, -0.051, 0])  # integers, floats, zero
     def test_ry_gradient(self, par, tol, dev):
         """Test that the gradient of the RY gate matches the exact analytic formula."""
@@ -669,6 +695,63 @@ class TestAdjointJacobianQNode:
         grad_F = grad_fn(*args)
 
         assert np.allclose(grad_A, grad_F, atol=tol, rtol=0)
+
+    @pytest.mark.skipif(
+        device_name != "lightning.qubit" or not ld._CPP_BINARY_AVAILABLE,
+        reason="N-controlled operations only implemented in lightning.qubit.",
+    )
+    @pytest.mark.parametrize(
+        "operation",
+        [
+            qml.PhaseShift,
+            qml.RX,
+            qml.RY,
+            qml.RZ,
+            qml.IsingXX,
+            qml.IsingXY,
+            qml.IsingYY,
+            qml.IsingZZ,
+            qml.SingleExcitation,
+            qml.SingleExcitationMinus,
+            qml.SingleExcitationPlus,
+            qml.DoubleExcitation,
+            qml.DoubleExcitationMinus,
+            qml.DoubleExcitationPlus,
+        ],
+    )
+    @pytest.mark.parametrize("n_qubits", range(2, 6))
+    @pytest.mark.parametrize("par", [-np.pi / 7, np.pi / 5, 2 * np.pi / 3])
+    def test_controlled_jacobian(self, par, n_qubits, operation, tol):
+        """Test that the jacobian of the controlled gate matches the parameter-shift formula."""
+        par = np.array([0.1234, par, 0.5678])
+        dev = qml.device("lightning.qubit", wires=n_qubits)
+        np.random.seed(1337)
+        init_state = np.random.rand(2**n_qubits) + 1.0j * np.random.rand(2**n_qubits)
+        init_state /= np.sqrt(np.dot(np.conj(init_state), init_state))
+
+        if operation.num_wires > n_qubits:
+            return
+
+        def circuit(p):
+            qml.StatePrep(init_state, wires=range(n_qubits))
+            qml.RX(p[0], 0)
+            qml.ctrl(
+                operation(p[1], wires=range(n_qubits - operation.num_wires, n_qubits)),
+                range(0, n_qubits - operation.num_wires),
+            )
+            qml.RY(p[2], 0)
+            return np.array([qml.expval(qml.PauliY(i)) for i in range(n_qubits)])
+
+        circ_ad = qml.QNode(circuit, dev, diff_method="adjoint")
+        circ_ps = qml.QNode(circuit, dev, diff_method="finite-diff")
+        jac_ad = np.array(qml.jacobian(circ_ad)(par))
+        jac_ps = np.array(qml.jacobian(circ_ps)(par))
+
+        # different methods must agree
+        assert jac_ad.size == n_qubits * 3
+        assert np.allclose(jac_ad.shape, [n_qubits, 3])
+        assert np.allclose(jac_ad.shape, jac_ps.shape)
+        assert np.allclose(jac_ad, jac_ps, atol=tol, rtol=0)
 
     thetas = np.linspace(-2 * np.pi, 2 * np.pi, 8)