+++ tests/test_gates.py
@@ -288,3 +288,174 @@ def test_qubit_unitary(n_wires, theta, phi, tol):
         circ = qml.QNode(circuit, dev)
         circ_def = qml.QNode(circuit, dev_def)
         assert np.allclose(circ(), circ_def(), tol)
+
+
+@pytest.mark.skipif(
+    device_name != "lightning.qubit",
+    reason="N-controlled operations only implemented in lightning.qubit.",
+)
+@pytest.mark.parametrize("control_value", [False, True])
+@pytest.mark.parametrize("n_qubits", list(range(2, 8)))
+def test_controlled_qubit_unitary(n_qubits, control_value, tol):
+    """Test that ControlledQubitUnitary is correctly applied to a state"""
+    dev_def = qml.device("default.qubit", wires=n_qubits)
+    dev = qml.device(device_name, wires=n_qubits)
+    threshold = 500
+    for n_wires in range(1, 5):
+        wire_lists = list(itertools.permutations(range(0, n_qubits), n_wires))
+        n_perms = len(wire_lists) * (n_wires) ** 2
+        if n_perms > threshold:
+            wire_lists = wire_lists[0 :: (n_perms // threshold)]
+        for all_wires in wire_lists:
+            for i in range(1, len(all_wires)):
+                control_wires = all_wires[0:i]
+                target_wires = all_wires[i:]
+                m = 2 ** len(target_wires)
+                U = np.random.rand(m, m) + 1.0j * np.random.rand(m, m)
+                U, _ = np.linalg.qr(U)
+                init_state = np.random.rand(2**n_qubits) + 1.0j * np.random.rand(2**n_qubits)
+                init_state /= np.sqrt(np.dot(np.conj(init_state), init_state))
+
+                def circuit():
+                    qml.StatePrep(init_state, wires=range(n_qubits))
+                    qml.ControlledQubitUnitary(
+                        U,
+                        control_wires=control_wires,
+                        wires=target_wires,
+                        control_values=[
+                            control_value or bool(i % 2) for i, _ in enumerate(control_wires)
+                        ],
+                    )
+                    return qml.state()
+
+                circ = qml.QNode(circuit, dev)
+                circ_def = qml.QNode(circuit, dev_def)
+                assert np.allclose(circ(), circ_def(), tol)
+
+
+@pytest.mark.skipif(
+    device_name != "lightning.qubit",
+    reason="N-controlled operations only implemented in lightning.qubit.",
+)
+@pytest.mark.parametrize(
+    "operation",
+    [
+        qml.PauliX,
+        qml.PauliY,
+        qml.PauliZ,
+        qml.Hadamard,
+        qml.S,
+        qml.T,
+        qml.PhaseShift,
+        qml.RX,
+        qml.RY,
+        qml.RZ,
+        qml.SWAP,
+        qml.IsingXX,
+        qml.IsingXY,
+        qml.IsingYY,
+        qml.IsingZZ,
+        qml.SingleExcitation,
+        qml.SingleExcitationMinus,
+        qml.SingleExcitationPlus,
+        qml.DoubleExcitation,
+        qml.DoubleExcitationMinus,
+        qml.DoubleExcitationPlus,
+    ],
+)
+@pytest.mark.parametrize("control_value", [False, True])
+@pytest.mark.parametrize("n_qubits", list(range(2, 8)))
+def test_controlled_qubit_gates(operation, n_qubits, control_value, tol):
+    """Test that multi-controlled gates are correctly applied to a state"""
+    dev_def = qml.device("default.qubit", wires=n_qubits)
+    dev = qml.device(device_name, wires=n_qubits)
+    threshold = 250
+    for n_wires in range(operation.num_wires + 1, operation.num_wires + 4):
+        wire_lists = list(itertools.permutations(range(0, n_qubits), n_wires))
+        n_perms = len(wire_lists) * n_wires
+        if n_perms > threshold:
+            wire_lists = wire_lists[0 :: (n_perms // threshold)]
+        for all_wires in wire_lists:
+            target_wires = all_wires[0 : operation.num_wires]
+            control_wires = all_wires[operation.num_wires :]
+            init_state = np.random.rand(2**n_qubits) + 1.0j * np.random.rand(2**n_qubits)
+            init_state /= np.sqrt(np.dot(np.conj(init_state), init_state))
+
+            def circuit():
+                qml.StatePrep(init_state, wires=range(n_qubits))
+                if operation.num_params == 0:
+                    qml.ctrl(
+                        operation(target_wires),
+                        control_wires,
+                        control_values=[
+                            control_value or bool(i % 2) for i, _ in enumerate(control_wires)
+                        ],
+                    )
+                else:
+                    qml.ctrl(
+                        operation(0.1234, target_wires),
+                        control_wires,
+                        control_values=[
+                            control_value or bool(i % 2) for i, _ in enumerate(control_wires)
+                        ],
+                    )
+                return qml.state()
+
+            circ = qml.QNode(circuit, dev)
+            circ_def = qml.QNode(circuit, dev_def)
+            assert np.allclose(circ(), circ_def(), tol)
+
+
+@pytest.mark.skipif(
+    device_name != "lightning.qubit",
+    reason="N-controlled operations only implemented in lightning.qubit.",
+)
+def test_controlled_qubit_unitary_from_op(tol):
+    n_qubits = 10
+    dev_def = qml.device("default.qubit", wires=n_qubits)
+    dev = qml.device(device_name, wires=n_qubits)
+
+    def circuit(x):
+        qml.ControlledQubitUnitary(
+            qml.QubitUnitary(qml.RX.compute_matrix(x), wires=5), control_wires=range(5)
+        )
+        return qml.expval(qml.PauliX(0))
+
+    circ = qml.QNode(circuit, dev)
+    circ_def = qml.QNode(circuit, dev_def)
+    par = 0.1234
+    assert np.allclose(circ(par), circ_def(par), tol)
+
+
+@pytest.mark.skipif(
+    device_name != "lightning.qubit",
+    reason="N-controlled operations only implemented in lightning.qubit.",
+)
+@pytest.mark.parametrize("control_wires", range(4))
+@pytest.mark.parametrize("target_wires", range(4))
+def test_cnot_controlled_qubit_unitary(control_wires, target_wires, tol):
+    """Test that ControlledQubitUnitary is correctly applied to a state"""
+    if control_wires == target_wires:
+        return
+    n_qubits = 4
+    control_wires = [control_wires]
+    target_wires = [target_wires]
+    dev = qml.device(device_name, wires=n_qubits)
+    wires = control_wires + target_wires
+    U = qml.matrix(qml.PauliX(target_wires))
+    init_state = np.random.rand(2**n_qubits) + 1.0j * np.random.rand(2**n_qubits)
+    init_state /= np.sqrt(np.dot(np.conj(init_state), init_state))
+
+    def circuit():
+        qml.StatePrep(init_state, wires=range(n_qubits))
+        qml.ControlledQubitUnitary(U, control_wires=control_wires, wires=target_wires)
+        return qml.state()
+
+    def cnot_circuit():
+        qml.StatePrep(init_state, wires=range(n_qubits))
+        qml.CNOT(wires=wires)
+        return qml.state()
+
+    circ = qml.QNode(circuit, dev)
+    circ_def = qml.QNode(cnot_circuit, dev)
+    assert np.allclose(circ(), circ_def(), tol)